Revert "DAOS-13882 test: Continue archiving on successful hosts. (#13110)"
This reverts commit a7578357c9016874c4011048f9180bdf599f7c2f.
