+++ src/tests/ftest/control/dmg_network_scan.py
@@ -4,9 +4,7 @@
   SPDX-License-Identifier: BSD-2-Clause-Patent
 """
 from apricot import TestWithServers
-from ClusterShell.NodeSet import NodeSet
-from exception_utils import CommandFailure
-from network_utils import SUPPORTED_PROVIDERS, NetworkDevice, get_network_information
+from network_utils import SUPPORTED_PROVIDERS, get_dmg_network_information, get_network_information
 
 
 class DmgNetworkScanTest(TestWithServers):
@@ -31,7 +29,7 @@ class DmgNetworkScanTest(TestWithServers):
         """
         server_provider = self.server_managers[0].get_config_value("provider")
         sys_info = []
-        for entry in get_network_information(self.log, self.hostlist_servers, SUPPORTED_PROVIDERS):
+        for entry in get_network_information(self.hostlist_servers, SUPPORTED_PROVIDERS):
             if server_provider in entry.provider:
                 entry.device = None
                 sys_info.append(entry)
@@ -45,37 +43,7 @@ class DmgNetworkScanTest(TestWithServers):
 
         """
         dmg = self.get_dmg_command()
-        return self.get_dmg_network_information(dmg.network_scan())
-
-    def get_dmg_network_information(self, dmg_network_scan):
-        """Get the network device information from the dmg network scan output.
-
-        Args:
-            dmg_network_scan (dict): the dmg network scan json command output
-
-        Raises:
-            CommandFailure: if there was an error processing the dmg network scan output
-
-        Returns:
-            list: a list of NetworkDevice objects identifying the network devices on each host
-
-        """
-        network_devices = []
-
-        try:
-            for host_fabric in dmg_network_scan["response"]["HostFabrics"].values():
-                for host in NodeSet(host_fabric["HostSet"].split(":")[0]):
-                    for interface in host_fabric["HostFabric"]["Interfaces"]:
-                        network_devices.append(
-                            NetworkDevice(
-                                host, interface["Device"], None, 1, interface["Provider"],
-                                interface["NumaNode"])
-                        )
-        except KeyError as error:
-            raise CommandFailure(
-                f"Error processing dmg network scan json output: {dmg_network_scan}") from error
-
-        return network_devices
+        return get_dmg_network_information(dmg.network_scan())
 
     def test_dmg_network_scan_basic(self):
         """JIRA ID: DAOS-2516
