+++ src/tests/ftest/process_core_files.py
@@ -5,7 +5,6 @@
 """
 import logging
 import os
-import re
 import sys
 from argparse import ArgumentParser
 from collections import defaultdict
@@ -13,7 +12,12 @@ from fnmatch import fnmatch
 
 # pylint: disable=import-error,no-name-in-module
 from util.logger_utils import get_console_handler
-from util.run_utils import RunException, find_command, run_local, run_remote
+from util.run_utils import RunException, find_command, run_local
+
+# Set up a logger for the console messages
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.DEBUG)
+logger.addHandler(get_console_handler("%(message)s", logging.DEBUG))
 
 # One of the dfuse tests intermittently creates core files which is known so make a special case
 # for that test.
@@ -22,47 +26,6 @@ CORE_FILES_IGNORE = {'./dfuse/daos_build.py': ('./conftest')}
 
 class CoreFileException(Exception):
     """Base exception for this module."""
-
-
-def get_core_file_pattern(log, hosts):
-    """Get the core file pattern information from the hosts if collecting core files.
-
-    Args:
-        log (Logger): logger for the messages produced by this method
-        clients (NodeSet): hosts designated for the client role in testing
-
-    Raises:
-        CoreFileException: if there was an error obtaining the core file pattern information
-
-    Returns:
-        dict: a dictionary containing the path and pattern for the core files per NodeSet
-    """
-    core_files = {}
-
-    # Determine the core file pattern being used by the hosts
-    command = "cat /proc/sys/kernel/core_pattern"
-    result = run_remote(log, hosts, command)
-
-    # Verify all the hosts have the same core file pattern
-    if not result.passed:
-        raise CoreFileException("Error obtaining the core file pattern")
-
-    # Get the path and pattern information from the core pattern
-    for data in result.output:
-        str_hosts = str(data.hosts)
-        try:
-            info = os.path.split(result.output[0].stdout[-1])
-        except (TypeError, IndexError) as error:
-            raise CoreFileException(
-                "Error obtaining the core file pattern and directory") from error
-        if not info[0]:
-            raise CoreFileException("Error obtaining the core file pattern directory")
-        core_files[str_hosts] = {"path": info[0], "pattern": re.sub(r"%[A-Za-z]", "*", info[1])}
-        log.info(
-            "Collecting any '%s' core files written to %s on %s",
-            core_files[str_hosts]["pattern"], core_files[str_hosts]["path"], str_hosts)
-
-    return core_files
 
 
 class CoreFileProcessing():
@@ -473,10 +436,4 @@ def main():
 
 
 if __name__ == "__main__":
-    # Set up a logger for the console messages
-    logger = logging.getLogger(__name__)
-    logger.setLevel(logging.DEBUG)
-    logger.addHandler(get_console_handler("%(message)s", logging.DEBUG))
     main()
-else:
-    logger = logging.getLogger()
