+++ src/tests/ftest/scripts/main.sh
@@ -192,4 +192,19 @@ else
     rc=0
 fi
 
+# daos_test uses cmocka framework which generates a set of xml of its own.
+# Post-processing the xml files here to put them in proper categories
+# for publishing in Jenkins
+TEST_DIRS=("daos_test" "checksum")
+
+for test_dir in "${TEST_DIRS[@]}"; do
+    COMP="FTEST_${test_dir}"
+    if [[ "${LAUNCH_OPT_ARGS}" == *"--repeat="* ]]; then
+        FILES=("${logs_prefix}/ftest/avocado/job-results/${test_dir}"/*/*/test-results/*/data/*.xml)
+    else
+        FILES=("${logs_prefix}/ftest/avocado/job-results/${test_dir}"/*/test-results/*/data/*.xml)
+    fi
+    ./scripts/post_process_xml.sh "${COMP}" "${FILES[@]}"
+done
+
 exit $rc
