+++ src/tests/ftest/util/network_utils.py
@@ -6,26 +6,13 @@
 import errno
 import os
 import re
+from logging import getLogger
 
 from ClusterShell.NodeSet import NodeSet
-# pylint: disable=import-error,no-name-in-module
-from util.run_utils import run_remote
+from exception_utils import CommandFailure
+from general_utils import display_task, run_pcmd, run_task
 
-SUPPORTED_PROVIDERS = (
-    "ofi+cxi",
-    "ofi+verbs;ofi_rxm",
-    "ucx+dc_x",
-    "ofi+tcp;ofi_rxm",
-    "ofi+opx"
-)
-PROVIDER_ALIAS = {
-    "ofi+verbs": "ofi+verbs;ofi_rxm",
-    "ofi+tcp": "ofi+tcp;ofi_rxm"
-}
-
-
-class NetworkException(Exception):
-    """Exception for network_utils methods."""
+SUPPORTED_PROVIDERS = ("ofi+sockets", "ofi+tcp;ofi_rxm", "ofi+verbs;ofi_rxm", "ucx+dc_x", "ofi+cxi")
 
 
 class NetworkDevice():
@@ -78,146 +65,158 @@ class NetworkDevice():
         return self.device
 
 
-def get_active_network_interfaces(logger, hosts, verbose=True):
+def get_active_network_interfaces(hosts, verbose=True):
     """Get all active network interfaces on the hosts.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
         hosts (NodeSet): hosts on which to find active interfaces
         verbose (bool, optional): display command details. Defaults to True.
 
     Returns:
         dict: a dictionary of interface keys and NodeSet values on which they were found
+
     """
     net_path = os.path.join(os.path.sep, "sys", "class", "net")
     operstate = os.path.join(net_path, "*", "operstate")
     command = " | ".join([f"grep -l 'up' {operstate}", "grep -Ev '/(lo|bond)/'", "sort"])
-    result = run_remote(logger, hosts, command, verbose)
+    task = run_task(hosts, command, verbose=verbose)
+    if verbose:
+        display_task(task)
 
     # Populate a dictionary of active interfaces with a NodSet of hosts on which it was found
     active_interfaces = {}
-    for data in result.output:
-        if not data.passed:
-            continue
-        for line in data.stdout:
+    for output, nodelist in task.iter_buffers():
+        output_lines = [line.decode("utf-8") for line in output]
+        nodeset = NodeSet.fromlist(nodelist)
+        for line in output_lines:
             try:
                 interface = line.split("/")[-2]
                 if interface not in active_interfaces:
                     active_interfaces[interface] = NodeSet()
-                active_interfaces[interface].update(data.hosts)
+                active_interfaces[interface].update(nodeset)
             except IndexError:
                 pass
 
     return active_interfaces
 
 
-def get_common_interfaces(logger, hosts, verbose=True):
+def get_common_interfaces(hosts, verbose=True):
     """Get the active interfaces that exists on all the hosts.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
         hosts (NodeSet): hosts on which to find active interfaces
         verbose (bool, optional): display command details. Defaults to True.
 
     Returns:
         list: a list of available NetworkDevice objects common to all hosts
+
     """
+    log = getLogger()
+
     # Find any active network interfaces on the server or client hosts
-    interfaces_per_host = get_active_network_interfaces(logger, hosts, verbose)
+    interfaces_per_host = get_active_network_interfaces(hosts, verbose)
 
     # From the active interface dictionary find all the interfaces that are common to all hosts
-    logger.info("Active network interfaces detected:")
+    log.info("Active network interfaces detected:")
     common_interfaces = []
     for interface, node_set in interfaces_per_host.items():
-        logger.info("  - %-8s on %s (Common=%s)", interface, node_set, node_set == hosts)
+        log.info("  - %-8s on %s (Common=%s)", interface, node_set, node_set == hosts)
         if node_set == hosts:
             common_interfaces.append(interface)
 
     return common_interfaces
 
 
-def get_interface_speeds(logger, hosts, interface, verbose=True):
+def get_interface_speeds(hosts, interface, verbose=True):
     """Get the speeds of this network interface on each host.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
         hosts (NodeSet): hosts on which to detect the interface speed
         interface (str): interface for which to obtain the speed
         verbose (bool, optional): display command details. Defaults to True.
 
     Returns:
         dict: a dictionary of interface speed keys and NodeSet values on which they were detected
+
     """
     net_path = os.path.join(os.path.sep, "sys", "class", "net")
     command = f"cat {os.path.join(net_path, interface, 'speed')}"
-    result = run_remote(logger, hosts, command, verbose)
+    task = run_task(hosts, command, verbose=verbose)
+    if verbose:
+        display_task(task)
 
     # Populate a dictionary of interface speeds with a NodSet of hosts on which it was detected
     interface_speeds = {}
-    for data in result.output:
-        if not data.passed:
-            continue
-        for line in data.stdout:
-            try:
-                speed = int(line.strip())
-            except IOError as io_error:
-                # KVM/Qemu/libvirt returns an EINVAL
-                if io_error.errno == errno.EINVAL:
-                    speed = 1000
-            except ValueError:
-                # Any line not containing a speed (integer)
-                continue
-            if speed not in interface_speeds:
-                interface_speeds[speed] = NodeSet()
-            interface_speeds[speed].update(data.hosts)
+    results = dict(task.iter_retcodes())
+    if 0 in results:
+        for output, nodelist in task.iter_buffers(results[0]):
+            output_lines = [line.decode("utf-8") for line in output]
+            nodeset = NodeSet.fromlist(nodelist)
+            for line in output_lines:
+                try:
+                    speed = int(line.strip())
+                except IOError as io_error:
+                    # KVM/Qemu/libvirt returns an EINVAL
+                    if io_error.errno == errno.EINVAL:
+                        speed = 1000
+                except ValueError:
+                    # Any line not containing a speed (integer)
+                    continue
+                if speed not in interface_speeds:
+                    interface_speeds[speed] = NodeSet()
+                interface_speeds[speed].update(nodeset)
 
     return interface_speeds
 
 
-def get_interface_numa_node(logger, hosts, interface, verbose=True):
+def get_interface_numa_node(hosts, interface, verbose=True):
     """Get the NUMA node ID of this network interface on each host.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
         hosts (NodeSet): hosts on which to detect the NUMA node
         interface (str): interface for which to obtain the NUMA node
         verbose (bool, optional): display command details. Defaults to True.
 
     Returns:
         dict: a dictionary of NUMA node ID keys and NodeSet values on which they were detected
+
     """
     net_path = os.path.join(os.path.sep, "sys", "class", "net")
     command = f"cat {os.path.join(net_path, interface, 'device', 'numa_node')}"
-    result = run_remote(logger, hosts, command, verbose)
+    task = run_task(hosts, command, verbose=verbose)
+    if verbose:
+        display_task(task)
 
     # Populate a dictionary of numa node IDs with a NodSet of hosts on which it was detected
     numa_nodes = {}
-    for data in result.output:
-        if not data.passed:
-            continue
-        for line in data.stdout:
-            try:
-                numa_node = int(line.strip())
-            except ValueError:
-                continue
-            if numa_node not in numa_nodes:
-                numa_nodes[numa_node] = NodeSet()
-            numa_nodes[numa_node].update(data.hosts)
+    results = dict(task.iter_retcodes())
+    if 0 in results:
+        for output, nodelist in task.iter_buffers(results[0]):
+            output_lines = [line.decode("utf-8") for line in output]
+            nodeset = NodeSet.fromlist(nodelist)
+            for line in output_lines:
+                try:
+                    numa_node = int(line.strip())
+                except ValueError:
+                    continue
+                if numa_node not in numa_nodes:
+                    numa_nodes[numa_node] = NodeSet()
+                numa_nodes[numa_node].update(nodeset)
 
     return numa_nodes
 
 
-def get_interface_device_name(logger, hosts, interface, verbose=True):
+def get_interface_device_name(hosts, interface, verbose=True):
     """Get the device name of this network interface on each host.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
         hosts (NodeSet): hosts on which to detect the device name
         interface (str): interface for which to obtain the device name
         verbose (bool, optional): display command details. Defaults to True
 
     Returns:
         dict: a dictionary of device name keys and NodeSet values on which they were detected
+
     """
     device_names = {}
     net_path = os.path.join(os.path.sep, "sys", "class", "net")
@@ -226,82 +225,29 @@ def get_interface_device_name(logger, hosts, interface, verbose=True):
     while not device_names and device_dirs:
         device_dir = device_dirs.pop(0)
         command = f"ls -1 {os.path.join(net_path, interface, 'device', device_dir)}"
-        result = run_remote(logger, hosts, command, verbose)
+        task = run_task(hosts, command, verbose=verbose)
+        if verbose:
+            display_task(task)
 
         # Populate a dictionary of IB names with a NodSet of hosts on which it was detected
-        for data in result.output:
-            if not data.passed:
-                continue
-            device_name_list = []
-            for line in data.stdout:
-                match = re.findall(r"([A-Za-z0-9;_+]+)", line)
-                if len(match) == 1:
-                    device_name_list.append(match[0])
-            if device_name_list:
-                device_names[",".join(device_name_list)] = NodeSet.fromlist(data.hosts)
+        results = dict(task.iter_retcodes())
+        if 0 in results:
+            for output, nodelist in task.iter_buffers(results[0]):
+                device_name_list = []
+                for line in output:
+                    match = re.findall(r"([A-Za-z0-9;_+]+)", line.decode("utf-8"))
+                    if len(match) == 1:
+                        device_name_list.append(match[0])
+                if device_name_list:
+                    device_names[",".join(device_name_list)] = NodeSet.fromlist(nodelist)
 
     return device_names
 
 
-def get_hg_info(logger, hosts, filter_provider=None, filter_device=None, verbose=True):
-    """Get the HG provider information from the specified hosts.
+def get_ofi_info(hosts, supported=None, verbose=True):
+    """Get the OFI provider information from the specified hosts.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
-        hosts (NodeSet): hosts from which to gather the information
-        filter_provider (list, optional): list of supported providers to filter by.
-            Defaults to None.
-        filter_device (list, optional): list of supported devices to filter by.
-            Defaults to None.
-        verbose (bool, optional): display command details. Defaults to True.
-
-    Returns:
-        dict: a dictionary of interface keys with a dictionary value of a comma-separated string of
-            providers key with a NodeSet value where the providers where detected.
-    """
-    command = "hg_info"
-    result = run_remote(logger, hosts, command, verbose=verbose, stderr=True)
-    providers = {}
-    if result.passed:
-        # Find all supported providers
-        for data in result.output:
-            if not data.stdout:
-                continue
-            # Skip over the header
-            without_header = re.findall(
-                r'^-+\n^.*\n^-+\n(.*)', '\n'.join(data.stdout), re.MULTILINE | re.DOTALL)[0]
-            # Convert:
-            # <Class>  <Protocol>  <Device>
-            # To {device: set(providers)}
-            device_providers = {}
-            class_protocol_device = re.findall(
-                r'(\S+) +([\S]+) +([\S]+)$', without_header, re.MULTILINE)
-            for _class, protocol, device in class_protocol_device:
-                if filter_device and device not in filter_device:
-                    continue
-                provider = f"{_class}+{protocol}"
-                if filter_provider and provider not in filter_provider:
-                    continue
-                if device not in device_providers:
-                    device_providers[device] = set()
-                device_providers[device].add(provider)
-
-            for device, provider_set in device_providers.items():
-                if device not in providers:
-                    providers[device] = {}
-                provider_key = ",".join(provider_set)
-                if provider_key not in providers[device]:
-                    providers[device][provider_key] = NodeSet()
-                providers[device][provider_key].update(data.hosts)
-
-    return providers
-
-
-def get_ucx_info(logger, hosts, supported=None, verbose=True):
-    """Get the UCX provider information from the specified hosts.
-
-    Args:
-        logger (Logger): logger for the messages produced by this method
         hosts (NodeSet): hosts from which to gather the information
         supported (list, optional): list of supported providers when if provided will limit the
             inclusion to only those providers specified. Defaults to None.
@@ -310,51 +256,121 @@ def get_ucx_info(logger, hosts, supported=None, verbose=True):
     Returns:
         dict: a dictionary of interface keys with a dictionary value of a comma-separated string of
             providers key with a NodeSet value where the providers where detected.
+
     """
-    result = run_remote(logger, hosts, "ucx_info -d", verbose)
+    task = run_task(hosts, "fi_info", verbose=verbose)
+    if verbose:
+        display_task(task)
 
     # Populate a dictionary of interfaces with a list of provider lists and NodSet of hosts on which
     # the providers were detected.
     providers = {}
-    for data in result.output:
-        if not data.passed:
-            continue
+    results = dict(task.iter_retcodes())
+    if 0 in results:
+        for output, nodelist in task.iter_buffers(results[0]):
+            output_lines = [line.decode("utf-8").rstrip(os.linesep) for line in output]
+            nodeset = NodeSet.fromlist(nodelist)
 
-        # Find all the transport, device, and type pairings. The ucx_info output reports these on
-        # separate lines so when processing the re matches ensure each device is preceded by a
-        # provider.
-        interface_providers = {}
-        matches = re.findall(r"(Transport|Device):\s+([A-Za-z0-9;_+]+)", "\n".join(data.stdout))
-        while matches:
-            transport = list(matches.pop(0))
-            if transport[0] == "Transport" and matches[0][0] == "Device":
-                transport.pop(0)
-                device = list(matches.pop(0))
-                device.pop(0)
+            # Find all the provider and domain pairings. The fi_info output reports these on
+            # separate lines when processing the re matches ensure each domain is preceded by a
+            # provider.
+            interface_providers = {}
+            data = re.findall(r"(provider|domain):\s+([A-Za-z0-9;_+]+)", "\n".join(output_lines))
+            while data:
+                provider = list(data.pop(0))
+                if provider[0] == "provider" and data[0][0] == "domain":
+                    provider.pop(0)
+                    domain = list(data.pop(0))
+                    domain.pop(0)
 
-                # A transport and device must be specified
-                if not transport or not device:
-                    continue
+                    # A provider and domain must be specified
+                    if not provider or not domain:
+                        continue
 
-                # Add 'ucx+' to the provider and replace 'mlx[0-9]' with 'x'
-                transport = [
-                    "+".join(["ucx", re.sub(r"mlx[0-9]+", "x", item)]) for item in transport]
+                    # Add 'ofi+' to the provider
+                    provider = ["+".join(["ofi", item]) for item in provider]
 
-                # Only include supported providers if a supported list is provided
-                if supported and transport[0] not in supported:
-                    continue
+                    # Only include supported providers if a supported list is provided
+                    if supported and provider[0] not in supported:
+                        continue
 
-                if device[0] not in interface_providers:
-                    interface_providers[device[0]] = set()
-                interface_providers[device[0]].update(transport)
+                    if domain[0] not in interface_providers:
+                        interface_providers[domain[0]] = set()
+                    interface_providers[domain[0]].update(provider)
 
-        for interface, provider_set in interface_providers.items():
-            if interface not in providers:
-                providers[interface] = {}
-            provider_key = ",".join(list(provider_set))
-            if provider_key not in providers[interface]:
-                providers[interface][provider_key] = NodeSet()
-            providers[interface][provider_key].update(data.hosts)
+            for interface, provider_set in interface_providers.items():
+                if interface not in providers:
+                    providers[interface] = {}
+                provider_key = ",".join(list(provider_set))
+                if provider_key not in providers[interface]:
+                    providers[interface][provider_key] = NodeSet()
+                providers[interface][provider_key].update(nodeset)
+
+    return providers
+
+
+def get_ucx_info(hosts, supported=None, verbose=True):
+    """Get the UCX provider information from the specified hosts.
+
+    Args:
+        hosts (NodeSet): hosts from which to gather the information
+        supported (list, optional): list of supported providers when if provided will limit the
+            inclusion to only those providers specified. Defaults to None.
+        verbose (bool, optional): display command details. Defaults to True.
+
+    Returns:
+        dict: a dictionary of interface keys with a dictionary value of a comma-separated string of
+            providers key with a NodeSet value where the providers where detected.
+
+    """
+    task = run_task(hosts, "ucx_info -d", verbose=verbose)
+    if verbose:
+        display_task(task)
+
+    # Populate a dictionary of interfaces with a list of provider lists and NodSet of hosts on which
+    # the providers were detected.
+    providers = {}
+    results = dict(task.iter_retcodes())
+    if 0 in results:
+        for output, nodelist in task.iter_buffers(results[0]):
+            output_lines = [line.decode("utf-8").rstrip(os.linesep) for line in output]
+            nodeset = NodeSet.fromlist(nodelist)
+
+            # Find all the transport, device, and type pairings. The ucx_info output reports these
+            # on separate lines so when processing the re matches ensure each device is preceded by
+            # a provider.
+            interface_providers = {}
+            data = re.findall(r"(Transport|Device):\s+([A-Za-z0-9;_+]+)", "\n".join(output_lines))
+            while data:
+                transport = list(data.pop(0))
+                if transport[0] == "Transport" and data[0][0] == "Device":
+                    transport.pop(0)
+                    device = list(data.pop(0))
+                    device.pop(0)
+
+                    # A transport and device must be specified
+                    if not transport or not device:
+                        continue
+
+                    # Add 'ucx+' to the provider and replace 'mlx[0-9]' with 'x'
+                    transport = [
+                        "+".join(["ucx", re.sub(r"mlx[0-9]+", "x", item)]) for item in transport]
+
+                    # Only include supported providers if a supported list is provided
+                    if supported and transport[0] not in supported:
+                        continue
+
+                    if device[0] not in interface_providers:
+                        interface_providers[device[0]] = set()
+                    interface_providers[device[0]].update(transport)
+
+            for interface, provider_set in interface_providers.items():
+                if interface not in providers:
+                    providers[interface] = {}
+                provider_key = ",".join(list(provider_set))
+                if provider_key not in providers[interface]:
+                    providers[interface][provider_key] = NodeSet()
+                providers[interface][provider_key].update(nodeset)
 
     return providers
 
@@ -364,7 +380,7 @@ def get_interface_providers(interface, provider_data):
 
     Args:
         interface (str): interface for which to obtain the InfiniBand name
-        provider_data (dict): output from get_hg_info() or get_ucx_info()
+        provider_data (dict): output from get_ofi_info() or get_ucx_info()
 
     Returns:
         dict: a dictionary of comma-separated strings of providers keys and NodeSet values on which
@@ -381,26 +397,24 @@ def get_interface_providers(interface, provider_data):
     return providers
 
 
-def get_fastest_interface(logger, hosts, verbose=True):
+def get_fastest_interface(hosts, verbose=True):
     """Get the fastest active interface common to all hosts.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
         hosts (NodeSet): hosts on which to find the fastest interfaces
         verbose (bool, optional): display command details. Defaults to True.
 
-    Raises:
-        NetworkException: if there is an error detecting the fastest active interface
-
     Returns:
         str: the fastest active interface common to all hosts specified
+
     """
-    common_interfaces = get_common_interfaces(logger, hosts, verbose)
+    log = getLogger()
+    common_interfaces = get_common_interfaces(hosts, verbose)
 
     # Find the speed of each common active interface in order to be able to choose the fastest
     interface_speeds = {}
     for interface in common_interfaces:
-        detected_speeds = get_interface_speeds(logger, hosts, interface, verbose)
+        detected_speeds = get_interface_speeds(hosts, interface, verbose)
         speed_list = []
         for speed, node_set in detected_speeds.items():
             if node_set == hosts:
@@ -409,31 +423,30 @@ def get_fastest_interface(logger, hosts, verbose=True):
         if speed_list:
             interface_speeds[interface] = min(speed_list)
 
-    logger.info("Active network interface speeds on %s:", hosts)
+    log.info("Active network interface speeds on %s:", hosts)
     available_interfaces = {}
     for interface in sorted(interface_speeds):
-        logger.info("  - %-8s (speed: %6s)", interface, interface_speeds[interface])
+        log.info("  - %-8s (speed: %6s)", interface, interface_speeds[interface])
 
         # Only include the first active interface (as determined by alphabetic sort) for each speed
         if interface_speeds[interface] not in available_interfaces:
             available_interfaces[interface_speeds[interface]] = interface
 
-    logger.info("Available interfaces on %s: %s", hosts, available_interfaces)
+    log.info("Available interfaces on %s: %s", hosts, available_interfaces)
     try:
         # Select the fastest active interface available by sorting the speed
         interface = available_interfaces[sorted(available_interfaces)[-1]]
-    except IndexError as error:
-        raise NetworkException("Error obtaining a default interface!") from error
+    except IndexError:
+        interface = None
 
-    logger.info("Fastest interface detected on %s: %s", hosts, interface)
+    log.info("Fastest interface detected on %s: %s", hosts, interface)
     return interface
 
 
-def get_common_provider(logger, hosts, interface, supported=None, verbose=True):
+def get_common_provider(hosts, interface, supported=None, verbose=True):
     """Get the list of providers supported by the interface on every host.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
         hosts (NodeSet): hosts on which to find the provider information
         interface (str): interface for which to obtain the providers
         supported (list, optional): list of supported providers when if provided will limit the
@@ -442,29 +455,33 @@ def get_common_provider(logger, hosts, interface, supported=None, verbose=True):
 
     Returns:
         list: a list of providers supported by the interface on every host
+
     """
-    hg_info = get_hg_info(logger, hosts, filter_provider=supported, verbose=verbose)
-    providers = get_interface_providers(interface, hg_info)
-    for dev_name in get_interface_device_name(logger, hosts, interface, verbose):
-        providers.update(get_interface_providers(dev_name, hg_info))
+    ofi_info = get_ofi_info(hosts, supported, verbose)
+    providers = get_interface_providers(interface, ofi_info)
+    for dev_name in get_interface_device_name(hosts, interface, verbose):
+        providers.update(get_interface_providers(dev_name, ofi_info))
+
+    if not supported or "ucx" in supported:
+        ucx_info = get_ucx_info(hosts, supported, verbose)
+        for dev_name in get_interface_device_name(hosts, interface, verbose):
+            for add_on in ([], ["1"]):
+                device = ":".join([dev_name] + add_on)
+                providers.update(get_interface_providers(device, ucx_info))
 
     # Only include the providers supported by the interface on all of the hosts
     common_providers = set()
-    if providers:
-        logger.debug("Detected providers:")
     for provider, node_set in providers.items():
-        logger.debug("  %4s: %s", provider, node_set)
         if node_set == hosts:
             common_providers.update(provider.split(","))
 
     return list(common_providers)
 
 
-def get_network_information(logger, hosts, supported=None, verbose=True):
+def get_network_information(hosts, supported=None, verbose=True):
     """Get the network device information on the hosts specified.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
         hosts (NodeSet): hosts on which to find the network information
         supported (list, optional): list of supported providers when if provided will limit the
             inclusion to only those providers specified. Defaults to None.
@@ -472,22 +489,23 @@ def get_network_information(logger, hosts, supported=None, verbose=True):
 
     Returns:
         list: a list of NetworkDevice objects identifying the network devices on each host
+
     """
     network_devices = []
 
-    ofi_info = get_hg_info(logger, hosts, filter_provider=supported, verbose=verbose)
-    ucx_info = get_ucx_info(logger, hosts, supported, verbose)
+    ofi_info = get_ofi_info(hosts, supported, verbose)
+    ucx_info = get_ucx_info(hosts, supported, verbose)
 
-    interfaces = get_active_network_interfaces(logger, hosts, verbose)
+    interfaces = get_active_network_interfaces(hosts, verbose)
     for host in hosts:
         for interface, node_set in interfaces.items():
             if host not in node_set:
                 continue
             kwargs = {"host": host, "name": interface, "port": 1}
             data_gather = {
-                "device": get_interface_device_name(logger, node_set, interface, verbose),
+                "device": get_interface_device_name(node_set, interface, verbose),
                 "provider": get_interface_providers(interface, ofi_info),
-                "numa": get_interface_numa_node(logger, node_set, interface, verbose),
+                "numa": get_interface_numa_node(node_set, interface, verbose),
             }
             for dev_name in data_gather["device"]:
                 for add_on in ([], ["1"]):
@@ -513,18 +531,48 @@ def get_network_information(logger, hosts, supported=None, verbose=True):
     return network_devices
 
 
-def update_network_interface(logger, interface, state, hosts, errors=None):
+def get_dmg_network_information(dmg_network_scan):
+    """Get the network device information from the dmg network scan output.
+
+    Args:
+        dmg_network_scan (dict): the dmg network scan json command output
+
+    Raises:
+        CommandFailure: if there was an error processing the dmg network scan output
+
+    Returns:
+        list: a list of NetworkDevice objects identifying the network devices on each host
+
+    """
+    network_devices = []
+
+    try:
+        for host_fabric in dmg_network_scan["response"]["HostFabrics"].values():
+            for host in NodeSet(host_fabric["HostSet"].split(":")[0]):
+                for interface in host_fabric["HostFabric"]["Interfaces"]:
+                    network_devices.append(
+                        NetworkDevice(
+                            host, interface["Device"], None, 1, interface["Provider"],
+                            interface["NumaNode"])
+                    )
+    except KeyError as error:
+        raise CommandFailure(
+            f"Error processing dmg network scan json output: {dmg_network_scan}") from error
+
+    return network_devices
+
+
+def update_network_interface(interface, state, hosts, errors=None):
     """Turn on or off the given network interface.
 
     Args:
-        logger (Logger): logger for the messages produced by this method
         interface (str): Interface name such as ib0.
         state (str): "up" or "down".
         hosts (NodeSet): Hosts on which the interface will be updated.
         errors (list): List to store the error message, if the command fails.
             Defaults to None.
     """
-    command = f"sudo ip link set {interface} {state}"
-    result = run_remote(logger, hosts, command)
-    if errors is not None and not result.passed:
-        errors.append(f"{command} didn't return succeed on {result.failed_hosts}")
+    command = "sudo ip link set {} {}".format(interface, state)
+    results = run_pcmd(hosts=hosts, command=command)
+    if errors is not None and results[0]["exit_status"] != 0:
+        errors.append(f"{command} didn't return 0!")
