+++ src/tests/ftest/util/results_utils.py
@@ -3,7 +3,7 @@
 
   SPDX-License-Identifier: BSD-2-Clause-Patent
 """
-import os
+
 import time
 from argparse import Namespace
 
@@ -29,6 +29,7 @@ class TestName():
 
         Returns:
             str: the test name including the uid, name, and variant
+
         """
         return "{}-{}{}".format(self.uid, self.name, self.variant)
 
@@ -76,6 +77,7 @@ class TestResult():
 
         Returns:
             object: the attribute value or default if not defined
+
         """
         return self.get(name, default)
 
@@ -89,6 +91,7 @@ class TestResult():
 
         Returns:
             bool: True if the attribute name has a value in this object; False otherwise
+
         """
         try:
             getattr(self, item)
@@ -105,6 +108,7 @@ class TestResult():
 
         Returns:
             object: the attribute value or default if not defined
+
         """
         try:
             return getattr(self, name, default)
@@ -124,98 +128,6 @@ class TestResult():
         self.time_end = time.time()
         # Increase the elapsed time by the delta between the last start call and this end call
         self.time_elapsed += self.time_end - self._time_split
-
-    def finish_test(self, logger, message, fail_class=None, exc_info=None):
-        """Mark the end of the test result with a status.
-
-        Args:
-            logger (Logger): logger for the messages produced by this method
-            message (str): exit message or reason for failure
-            fail_class (str, optional): failure category.
-            exc_info (OptExcInfo, optional): return value from sys.exc_info().
-        """
-        if fail_class is None:
-            self.pass_test(logger, message)
-        else:
-            self.fail_test(logger, fail_class, message, exc_info)
-        self.end()
-
-    def pass_test(self, logger, message=None):
-        """Set the test result as passed.
-
-        Args:
-            logger (Logger): logger for the messages produced by this method
-            message (str, optional): explanation of test passing. Defaults to None.
-        """
-        if message is not None:
-            logger.debug(message)
-        self.__set_test_status(logger, TestResult.PASS, None, None)
-
-    def warn_test(self, logger, fail_class, fail_reason, exc_info=None):
-        """Set the test result as warned.
-
-        Args:
-            logger (Logger): logger for the messages produced by this method
-            fail_class (str): failure category.
-            fail_reason (str): failure description.
-            exc_info (OptExcInfo, optional): return value from sys.exc_info(). Defaults to None.
-        """
-        logger.warning(fail_reason)
-        self.__set_test_status(logger, TestResult.WARN, fail_class, fail_reason, exc_info)
-
-    def fail_test(self, logger, fail_class, fail_reason, exc_info=None):
-        """Set the test result as failed.
-
-        Args:
-            logger (Logger): logger for the messages produced by this method
-            fail_class (str): failure category.
-            fail_reason (str): failure description.
-            exc_info (OptExcInfo, optional): return value from sys.exc_info(). Defaults to None.
-        """
-        logger.error(fail_reason)
-        self.__set_test_status(logger, TestResult.ERROR, fail_class, fail_reason, exc_info)
-
-    def __set_test_status(self, logger, status, fail_class, fail_reason, exc_info=None):
-        """Set the test result.
-
-        Args:
-            logger (Logger): logger for the messages produced by this method
-            status (str): TestResult status to set.
-            fail_class (str): failure category.
-            fail_reason (str): failure description.
-            exc_info (OptExcInfo, optional): return value from sys.exc_info(). Defaults to None.
-        """
-        if exc_info is not None:
-            logger.debug("Stacktrace", exc_info=True)
-
-        if status == TestResult.PASS:
-            # Do not override a possible WARN status
-            if self.status is None:
-                self.status = status
-            return
-
-        if self.fail_count == 0 or self.status == TestResult.WARN and status == TestResult.ERROR:
-            # Update the test result with the information about the first ERROR.
-            # Elevate status from WARN to ERROR if WARN came first.
-            self.status = status
-            self.fail_class = fail_class
-            self.fail_reason = fail_reason
-            if exc_info is not None:
-                try:
-                    # pylint: disable=import-outside-toplevel
-                    from avocado.utils.stacktrace import prepare_exc_info
-                    self.traceback = prepare_exc_info(exc_info)
-                except Exception:       # pylint: disable=broad-except
-                    pass
-
-        if self.fail_count > 0:
-            # Additional ERROR/WARN only update the test result fail reason with a fail counter
-            plural = "s" if self.fail_count > 1 else ""
-            fail_reason = self.fail_reason.split(" (+")[0:1]
-            fail_reason.append(f"{self.fail_count} other failure{plural})")
-            self.fail_reason = " (+".join(fail_reason)
-
-        self.fail_count += 1
 
 
 class Results():
@@ -275,6 +187,7 @@ class Results():
 
         Returns:
             int: total number of tests with errors
+
         """
         test_status = [str(test.status) for test in self.tests]
         return test_status.count(TestResult.ERROR)
@@ -285,6 +198,7 @@ class Results():
 
         Returns:
             int: total number of interrupted tests
+
         """
         test_status = [str(test.status) for test in self.tests]
         return test_status.count(TestResult.INTERRUPT)
@@ -295,6 +209,7 @@ class Results():
 
         Returns:
             int: total number of failed tests
+
         """
         test_status = [str(test.status) for test in self.tests]
         return test_status.count(TestResult.FAIL)
@@ -305,6 +220,7 @@ class Results():
 
         Returns:
             int: total number of skipped tests
+
         """
         test_status = [str(test.status) for test in self.tests]
         return test_status.count(TestResult.SKIP)
@@ -315,6 +231,7 @@ class Results():
 
         Returns:
             int: total number of cancelled tests
+
         """
         test_status = [str(test.status) for test in self.tests]
         return test_status.count(TestResult.CANCEL)
@@ -325,6 +242,7 @@ class Results():
 
         Returns:
             int: total number of tests that were not skipped or cancelled
+
         """
         return self.tests_total - self.skipped - self.cancelled
 
@@ -334,6 +252,7 @@ class Results():
 
         Returns:
             int: total number of tests that passed or warned
+
         """
         return self.passed + self.warned
 
@@ -343,24 +262,11 @@ class Results():
 
         Returns:
             float: pass rate of all the tests
+
         """
         if not self.completed:
             return 0.0
         return 100 * (float(self.succeeded) / float(self.completed))
-
-    def add_test(self, class_name, test_name, log_file):
-        """Add a new test result.
-
-        Args:
-            class_name (str): the test class name
-            test_name (TestName): the test uid, name, and variant
-            log_file (str): the log file for a single test
-
-        Returns:
-            TestResult: the test result for this test
-        """
-        self.tests.append(TestResult(class_name, test_name, log_file, self.logfile))
-        return self.tests[-1]
 
 
 class Job():
@@ -413,24 +319,6 @@ class Job():
         # If set to either 'RUNNING', 'ERROR', or 'FAIL' an html result will not be generated
         self.status = "COMPLETE"
 
-    def generate_results(self, logger, results):
-        """Generate the results.xml and results.html for this job.
-
-        Args:
-            logger (Logger): logger for the messages produced by this method
-            results (Results): the test results to use to generate the files
-        """
-        for key, create_method in {"results.xml": create_xml, "results.html": create_html}.items():
-            output = os.path.join(self.logdir, key)
-            try:
-                logger.debug("Creating %s: %s", key, output)
-                create_method(self, results)
-            except Exception as error:      # pylint: disable=broad-except
-                logger.error("Unable to create %s file: %s", key, error)
-            else:
-                if not os.path.exists(output):
-                    logger.error("The %s does not exist: %s", key, output)
-
 
 def sanitize_results(results):
     """Ensure each test status is set and all failure attributes are set for test failures.
@@ -481,100 +369,3 @@ def create_html(job, results):
     from avocado_result_html import HTMLResult
     result_html = HTMLResult()
     result_html.render(sanitize_results(results), job)
-
-
-class LaunchTestName():
-    """Define a launch.py test name compatible with avocado's result render classes."""
-
-    def __init__(self, name, order, repeat):
-        """Initialize a LaunchTestName object.
-
-        Args:
-            name (str): test name
-            order (int): order in which this test is executed
-            repeat (int): repeat count for this test
-        """
-        self.name = name
-        self.order = order
-        self.repeat = repeat
-
-    def __str__(self):
-        """Get the test name as a string.
-
-        Returns:
-            str: combination of the order and name
-        """
-        if self.repeat > 0:
-            return f"{self.uid}-{self.name}{self.variant}"
-        return f"{self.uid}-{self.name}"
-
-    def __getitem__(self, name, default=None):
-        """Get the value of the attribute name.
-
-        Args:
-            name (str): name of the class attribute to get
-            default (object, optional): value to return if name is not defined. Defaults to None.
-
-        Returns:
-            object: the attribute value or default if not defined
-        """
-        return self.get(name, default)
-
-    def get(self, name, default=None):
-        """Get the value of the attribute name.
-
-        Args:
-            name (str): name of the class attribute to get
-            default (object, optional): value to return if name is not defined. Defaults to None.
-
-        Returns:
-            object: the attribute value or default if not defined
-        """
-        try:
-            return getattr(self, name, default)
-        except TypeError:
-            return default
-
-    @property
-    def order_str(self):
-        """Get the string representation of the order count.
-
-        Returns:
-            str: the order count as a string
-        """
-        return f"{self.order:02}"
-
-    @property
-    def repeat_str(self):
-        """Get the string representation of the repeat count.
-
-        Returns:
-            str: the repeat count as a string
-        """
-        return f"repeat{self.repeat:03}"
-
-    @property
-    def uid(self):
-        """Get the test order to use as the test uid for xml/html results.
-
-        Returns:
-            str: the test uid (order)
-        """
-        return self.order_str
-
-    @property
-    def variant(self):
-        """Get the test repeat count as the test variant for xml/html results.
-
-        Returns:
-            str: the test variant (repeat)
-        """
-        return f";{self.repeat_str}"
-
-    def copy(self):
-        """Create a copy of this object.
-
-        Returns:
-            LaunchTestName: a copy of this LaunchTestName object
-        """
-        return LaunchTestName(self.name, self.order, self.repeat)
