+++ utils/cq/words.dict
@@ -235,7 +235,6 @@ junit
 kds
 kv
 kwargs
-launchable
 ldflags
 len
 libaio
@@ -442,7 +441,6 @@ testdata
 testFile
 testname
 testrunner
-testsuites
 tgt
 timestamp
 timestamps
